% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_grayscale.R
\name{filter_grayscale}
\alias{filter_grayscale}
\title{Vectorized Grayscale Filter}
\usage{
filter_grayscale(your_image)
}
\arguments{
\item{your_image}{Any image object loaded into R of common type or pathname to your image}
}
\value{
A filtered image with grayscale values applied. (using image_read from library(magick))
}
\description{
Converts any image of common format (eg .jpg .png. etc.) or pathname to your image file
Applies grayscale filter according to 0.299 x R,  0.587 x G,  and 0.114 x B
uses the magick package and outputs image file using image_read function.
}
\examples{
\dontrun{
# Load example image from the package
test_image_path <- system.file("extdata", "test.jpg", package = "VectorizedColorFilter")
grayscale_image <- filter_grayscale(test_image_path)

# Apply filter to an image object
img <- magick::image_read(test_image_path)
grayscale_image <- filter_grayscale(img)
}
}
