% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_twocolor.R
\name{filter_twocolor}
\alias{filter_twocolor}
\title{Vectorized Two-Color Filter}
\usage{
filter_twocolor(your_image, cutoff = 97, color1 = "pink", color2 = "blue")
}
\arguments{
\item{your_image}{An image as either a file path or image of standard format (.jpg .png etc.).}

\item{cutoff}{default = 97 (works well with included example) A numeric decimal value for the grayscale cutoff to apply two color filter.}

\item{color1}{default "pink" A  rgb triplet vector to select color 1 in twocolor image filter, R standard colors within "" (eg. "pink" )also works}

\item{color2}{default "blue" A rgb triplet vector to select color 2 in twocolor image filter, R standard colors within "" also works}
}
\value{
A filtered bitmap with color_1 and color_2 values applied.
}
\description{
Vectorized Two-Color Filter
}
