% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_grayscale.R
\name{filter_grayscale}
\alias{filter_grayscale}
\title{Vectorized Grayscale Filter}
\usage{
filter_grayscale(your_image)
}
\arguments{
\item{your_image}{Any image object loaded into R of common type or pathname to your image}
}
\value{
A filtered image with grayscale values applied. (using image_read from library(magick))
}
\description{
Converts any image of common format (eg .jpg .png. etc.) or pathname to your image file
Applies grayscale filter according to 0.299 x R,  0.587 x G,  and 0.114 x B
uses the magick package and outputs image file using image_read function.
}
\examples{
library(magick)
# Example using an image file
grayscale_image <- filter_grayscale("path/to/image.jpg")
# Example using an image object
img <- image_read("path/to/image.jpg")
grayscale_image <- filter_grayscale(img)
}
