% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_grayscale.R
\name{filter_grayscale}
\alias{filter_grayscale}
\title{Vectorized Grayscale Filter using
Converts any image of common format (eg .jpg .png. etc.) or pathname to your image file
Applies 0.299 * R,  0.587 * G,  and 0.114 * B}
\usage{
filter_grayscale(your_image)
}
\arguments{
\item{your_image}{Any image object loaded into R of common type or pathname to your image}
}
\value{
A filtered bitmap with grayscale values applied.
}
\description{
Vectorized Grayscale Filter using
Converts any image of common format (eg .jpg .png. etc.) or pathname to your image file
Applies 0.299 * R,  0.587 * G,  and 0.114 * B
}
