% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_twocolor.R
\name{filter_twocolor}
\alias{filter_twocolor}
\title{Vectorized Two-Color Filter}
\usage{
filter_twocolor(your_image, cutoff = 97, color1 = "pink", color2 = "blue")
}
\arguments{
\item{your_image}{An image as either a file path or image of standard format (.jpg .png etc.).}

\item{cutoff}{the default works well with supplied picture A numeric decimal value for the grayscale cutoff to apply two color filter.}

\item{color1}{A  rgb triplet vector to select color 1 in twocolor image filter, R standard colors within "" (eg. "pink" )also works}

\item{color2}{A rgb triplet vector to select color 2 in twocolor image filter, R standard colors within "" also works}
}
\value{
A filtered bitmap with color_1 and color_2 values applied.
}
\description{
Converts any image of common format (eg .jpg .png. etc.) or pathname to your image file
Applies grayscale filter according to 0.299 x R,  0.587 x G,  and 0.114 x B

Then applies a cutoff to the numerical gray scale value (0-255) and two colors to the upper and lower range
uses the magick package and outputs image file using image_read function.
}
\examples{
# Load example image using system.file() for portability
test_image_path <- system.file("extdata", "test.jpg", package = "VectorizedColorFilter")
twocolor_image <- filter_twocolor(test_image_path)  # Apply grayscale filter
twocolor_image
# Example using an image object
img <- magick::image_read(test_image_path)
twocolor_image <- filter_twocolor(img)
print (twocolor_image)
}
